Properly implement reliable datagrams
In previous versions, the receiver would ping senders once, upon which
the senders would send all available data in individual datagrams.
Turns out this corrupts the internal stack of the reliable datagram
manager. Reliable datagrams must always be paired in a single send
and a single response, which is what is implemented now.